Fix CPU parsing code
The CPU parsing code was counting the number of directories containing
"cpu" in the /sys/devices/system/cpu directory.  Unfortunately, it does
not follow that the number of CPUs matches that - since there may be
the presence of the "cpufreq" directory when cpufreq support is enabled.

Instead, count the number of CPUs found in /sys/bus/cpu/devices, which
are symlinks to the CPU directories found in the /sys/devices path.

Signed-off-by: Russell King <rmk@arm.linux.org.uk>